start/stop geotrigerring
