+++ app/src/main/res/layout/activity_main.xml
@@ -8,8 +8,8 @@
 
 
     <Button
-        android:id="@+id/bStopSDK"
+        android:id="@+id/bStartStopSDK"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
-        android:text="Stop SDK" />
+        android:text="@string/start_sdk" />
 </LinearLayout>
